refactor(agent-chat): simplify subagent transcript correlation
Route delayed subagent session links through canonical subagent part emission.

Simplify hydrated transcript merging for identified and unidentified subagent records.

Split subagent regression coverage out of oversized test files.
