+++ packages/adapters-opencode-sdk/src/event-stream.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "bun:test";
-import type { Event, OpencodeClient } from "@opencode-ai/sdk/v2/client";
+import type { Event } from "@opencode-ai/sdk/v2/client";
 import type { AgentEvent, AgentModelSelection, AgentUserMessagePart } from "@openducktor/core";
 import {
   isRelevantSubscriberEvent,
@@ -10,9 +10,14 @@ import {
   type EventStreamRuntime,
   flushPendingSubagentInputEventsForSession,
   readParentExternalSessionId,
-  type SubagentSessionLink,
 } from "./event-stream/shared";
-import type { SessionInput, SessionRecord } from "./types";
+import {
+  makeClientWithEvents,
+  makeSessionInput,
+  makeSessionRecord,
+  runEventStream,
+  runEventStreamWithSession,
+} from "./event-stream.test-support";
 import {
   buildQueuedRequestAttachmentIdentitySignature,
   buildQueuedRequestSignature,
@@ -40,99 +45,9 @@ const PDF_ATTACHMENT_DISPLAY_PART = {
   },
 };
 
-const makeClientWithEvents = (events: Event[]): OpencodeClient => {
-  return {
-    global: {
-      event: async () => {
-        async function* iterator(): AsyncGenerator<{ directory: string; payload: Event }> {
-          for (const event of events) {
-            const directory =
-              (event as Event & { properties?: { directory?: string } }).properties?.directory ??
-              "/repo";
-            yield { directory, payload: event };
-          }
-        }
-        return { stream: iterator() };
-      },
-    },
-  } as unknown as OpencodeClient;
-};
-
-const makeSessionInput = (): SessionInput => ({
-  repoPath: "/repo",
-  runtimeKind: "opencode",
-  workingDirectory: "/repo",
-  taskId: "task-1",
-  role: "spec",
-  systemPrompt: "System prompt",
-});
-
-const makeSessionRecord = (client: OpencodeClient): SessionRecord => ({
-  summary: {
-    externalSessionId: "external-session-1",
-    role: "spec",
-    startedAt: "2026-02-22T12:00:00.000Z",
-    status: "running",
-  },
-  input: makeSessionInput(),
-  client,
-  externalSessionId: "external-session-1",
-  eventTransportKey: "http://127.0.0.1:12345",
-  hasIdleSinceActivity: false,
-  activeAssistantMessageId: null,
-  completedAssistantMessageIds: new Set<string>(),
-  emittedAssistantMessageIds: new Set<string>(),
-  emittedUserMessageSignatures: new Map<string, string>(),
-  emittedUserMessageStates: new Map(),
-  pendingQueuedUserMessages: [],
-  partsById: new Map(),
-  messageRoleById: new Map(),
-  messageMetadataById: new Map(),
-  pendingDeltasByPartId: new Map(),
-  subagentCorrelationKeyByPartId: new Map(),
-  subagentCorrelationKeyByExternalSessionId: new Map(),
-  pendingSubagentCorrelationKeysBySignature: new Map(),
-  pendingSubagentCorrelationKeys: [],
-  pendingSubagentSessionsByExternalSessionId: new Map(),
-  pendingSubagentPartEmissionsByExternalSessionId: new Map(),
-  pendingSubagentInputEventsByExternalSessionId: new Map(),
-});
-
 const buildQueuedSignature = (message: string, model?: AgentModelSelection | null): string => {
   const parts: AgentUserMessagePart[] = [{ kind: "text", text: message }];
   return buildQueuedRequestSignature(parts, model ?? undefined);
-};
-
-const runEventStreamWithSession = async (
-  events: Event[],
-  configureSession?: (sessionRecord: SessionRecord) => void,
-  resolveSubagentSessionLink?: (childExternalSessionId: string) => SubagentSessionLink | undefined,
-): Promise<{ emitted: AgentEvent[]; sessionRecord: SessionRecord }> => {
-  const client = makeClientWithEvents(events);
-  const emitted: AgentEvent[] = [];
-  const sessionRecord = makeSessionRecord(client);
-  configureSession?.(sessionRecord);
-
-  await subscribeOpencodeEvents({
-    context: {
-      externalSessionId: "external-session-1",
-      input: makeSessionInput(),
-    },
-    client,
-    controller: new AbortController(),
-    now: () => "2026-02-22T12:00:00.000Z",
-    emit: (_sessionId, event) => {
-      emitted.push(event);
-    },
-    getSession: () => sessionRecord,
-    ...(resolveSubagentSessionLink ? { resolveSubagentSessionLink } : {}),
-  });
-
-  return { emitted, sessionRecord };
-};
-
-const runEventStream = async (events: Event[]): Promise<AgentEvent[]> => {
-  return (await runEventStreamWithSession(events)).emitted;
 };
 
 test("readParentExternalSessionId accepts OpenCode parent id spellings", () => {
@@ -385,30 +300,6 @@ const makeAssistantSubtaskPartUpdatedEvent = (input: {
         agent: input.agent,
         prompt: input.prompt,
         description: input.description,
-      },
-    },
-  }) as unknown as Event;
-
-const makeChildSessionCreatedEvent = (input: {
-  childSessionId: string;
-  parentExternalSessionId: string;
-}): Event =>
-  ({
-    type: "session.created",
-    properties: {
-      sessionID: input.childSessionId,
-      info: {
-        id: input.childSessionId,
-        slug: input.childSessionId,
-        projectID: "project-1",
-        directory: "/repo",
-        title: input.childSessionId,
-        version: "1",
-        parentID: input.parentExternalSessionId,
-        time: {
-          created: Date.parse("2026-02-22T12:00:10.000Z"),
-          updated: Date.parse("2026-02-22T12:00:10.000Z"),
-        },
       },
     },
   }) as unknown as Event;
@@ -1731,575 +1622,6 @@ describe("event-stream", () => {
       throw new Error("Expected assistant text part event");
     }
     expect(partEvents[0].part.text).toBe("Late role text");
-  });
-
-  test("binds same-turn sibling subagents to child sessions without fragmenting their cards", async () => {
-    const emitted = await runEventStream([
-      assistantRoleEvent("assistant-subagent-collision"),
-      makeAssistantSubtaskPartUpdatedEvent({
-        messageId: "assistant-subagent-collision",
-        partId: "subtask-a",
-        agent: "build",
-        prompt: "Inspect repo",
-        description: "Starting A",
-      }),
-      makeAssistantSubtaskPartUpdatedEvent({
-        messageId: "assistant-subagent-collision",
-        partId: "subtask-b",
-        agent: "build",
-        prompt: "Inspect repo",
-        description: "Starting B",
-      }),
-      makeChildSessionCreatedEvent({
-        childSessionId: "child-a",
-        parentExternalSessionId: "external-session-1",
-      }),
-      makeChildSessionCreatedEvent({
-        childSessionId: "child-b",
-        parentExternalSessionId: "external-session-1",
-      }),
-      {
-        type: "message.part.updated",
-        properties: {
-          part: {
-            id: "tool-a",
-            sessionID: "external-session-1",
-            messageID: "assistant-subagent-collision",
-            callID: "call-a",
-            type: "tool",
-            tool: "delegate",
-            state: {
-              status: "completed",
-              input: {
-                agent: "build",
-                prompt: "Inspect repo",
-              },
-              output: {
-                result: "Finished A",
-                externalSessionId: "child-a",
-              },
-            },
-          },
-        },
-      } as unknown as Event,
-      {
-        type: "message.part.updated",
-        properties: {
-          part: {
-            id: "tool-b",
-            sessionID: "external-session-1",
-            messageID: "assistant-subagent-collision",
-            callID: "call-b",
-            type: "tool",
-            tool: "delegate",
-            state: {
-              status: "completed",
-              input: {
-                agent: "build",
-                prompt: "Inspect repo",
-              },
-              output: {
-                result: "Finished B",
-                externalSessionId: "child-b",
-              },
-            },
-          },
-        },
-      } as unknown as Event,
-    ]);
-
-    const assistantPartEvents = emitted.filter(
-      (event): event is Extract<AgentEvent, { type: "assistant_part" }> =>
-        event.type === "assistant_part",
-    );
-    const subagentParts = assistantPartEvents
-      .map((event): Extract<AgentEvent, { type: "assistant_part" }>["part"] => event.part)
-      .filter(
-        (
-          part,
-        ): part is Extract<
-          Extract<AgentEvent, { type: "assistant_part" }>["part"],
-          { kind: "subagent" }
-        > => part.kind === "subagent",
-      );
-
-    expect(subagentParts).toHaveLength(6);
-
-    const runningParts = subagentParts.filter((part) => part.status === "running");
-    const completedParts = subagentParts.filter((part) => part.status === "completed");
-
-    expect(runningParts).toHaveLength(4);
-    expect(completedParts).toHaveLength(2);
-
-    const runningKeys = runningParts.map((part) => part.correlationKey);
-    expect(new Set(runningKeys).size).toBe(2);
-    expect(runningKeys).toEqual([
-      "part:assistant-subagent-collision:subtask-a",
-      "part:assistant-subagent-collision:subtask-b",
-      "part:assistant-subagent-collision:subtask-a",
-      "part:assistant-subagent-collision:subtask-b",
-    ]);
-    expect(runningParts.map((part) => part.externalSessionId)).toEqual([
-      undefined,
-      undefined,
-      "child-a",
-      "child-b",
-    ]);
-
-    expect(completedParts.map((part) => part.correlationKey)).toEqual([
-      "part:assistant-subagent-collision:subtask-a",
-      "part:assistant-subagent-collision:subtask-b",
-    ]);
-    expect(completedParts.map((part) => part.externalSessionId)).toEqual(["child-a", "child-b"]);
-    expect(completedParts.map((part) => part.description)).toEqual(["Finished A", "Finished B"]);
-  });
-
-  test("keeps ambiguous sibling completions bound to their original cards even when they finish out of order", async () => {
-    const { emitted } = await runEventStreamWithSession([
-      assistantRoleEvent("assistant-subagent-out-of-order"),
-      makeAssistantSubtaskPartUpdatedEvent({
-        messageId: "assistant-subagent-out-of-order",
-        partId: "subtask-a",
-        agent: "build",
-        prompt: "Inspect repo",
-        description: "Starting A",
-      }),
-      makeAssistantSubtaskPartUpdatedEvent({
-        messageId: "assistant-subagent-out-of-order",
-        partId: "subtask-b",
-        agent: "build",
-        prompt: "Inspect repo",
-        description: "Starting B",
-      }),
-      makeChildSessionCreatedEvent({
-        childSessionId: "child-a",
-        parentExternalSessionId: "external-session-1",
-      }),
-      makeChildSessionCreatedEvent({
-        childSessionId: "child-b",
-        parentExternalSessionId: "external-session-1",
-      }),
-      {
-        type: "message.part.updated",
-        properties: {
-          part: {
-            id: "tool-b",
-            sessionID: "external-session-1",
-            messageID: "assistant-subagent-out-of-order",
-            callID: "call-b",
-            type: "tool",
-            tool: "delegate",
-            state: {
-              status: "completed",
-              input: {
-                agent: "build",
-                prompt: "Inspect repo",
-              },
-              output: {
-                result: "Finished B",
-                externalSessionId: "child-b",
-              },
-            },
-          },
-        },
-      } as unknown as Event,
-      {
-        type: "message.part.updated",
-        properties: {
-          part: {
-            id: "tool-a",
-            sessionID: "external-session-1",
-            messageID: "assistant-subagent-out-of-order",
-            callID: "call-a",
-            type: "tool",
-            tool: "delegate",
-            state: {
-              status: "completed",
-              input: {
-                agent: "build",
-                prompt: "Inspect repo",
-              },
-              output: {
-                result: "Finished A",
-                externalSessionId: "child-a",
-              },
-            },
-          },
-        },
-      } as unknown as Event,
-    ]);
-
-    const assistantPartEvents = emitted.filter(
-      (event): event is Extract<AgentEvent, { type: "assistant_part" }> =>
-        event.type === "assistant_part",
-    );
-    const subagentParts = assistantPartEvents
-      .map((event): Extract<AgentEvent, { type: "assistant_part" }>["part"] => event.part)
-      .filter(
-        (
-          part,
-        ): part is Extract<
-          Extract<AgentEvent, { type: "assistant_part" }>["part"],
-          { kind: "subagent" }
-        > => part.kind === "subagent",
-      );
-
-    expect(subagentParts).toHaveLength(6);
-
-    const runningParts = subagentParts.filter((part) => part.status === "running");
-    const completedParts = subagentParts.filter((part) => part.status === "completed");
-
-    expect(runningParts.map((part) => part.correlationKey)).toEqual([
-      "part:assistant-subagent-out-of-order:subtask-a",
-      "part:assistant-subagent-out-of-order:subtask-b",
-      "part:assistant-subagent-out-of-order:subtask-a",
-      "part:assistant-subagent-out-of-order:subtask-b",
-    ]);
-    expect(runningParts.map((part) => part.externalSessionId)).toEqual([
-      undefined,
-      undefined,
-      "child-a",
-      "child-b",
-    ]);
-    expect(completedParts.map((part) => part.correlationKey)).toEqual([
-      "part:assistant-subagent-out-of-order:subtask-b",
-      "part:assistant-subagent-out-of-order:subtask-a",
-    ]);
-    expect(completedParts.map((part) => part.externalSessionId)).toEqual(["child-b", "child-a"]);
-    expect(completedParts.map((part) => part.description)).toEqual(["Finished B", "Finished A"]);
-  });
-
-  test("binds running task tool updates with metadata session ids back to their spawned card", async () => {
-    const emitted = await runEventStream([
-      assistantRoleEvent("assistant-task-tool-running"),
-      makeAssistantSubtaskPartUpdatedEvent({
-        messageId: "assistant-task-tool-running",
-        partId: "subtask-a",
-        agent: "build",
-        prompt: "Inspect repo",
-        description: "Starting A",
-      }),
-      {
-        type: "message.part.updated",
-        properties: {
-          part: {
-            id: "tool-a",
-            sessionID: "external-session-1",
-            messageID: "assistant-task-tool-running",
-            callID: "call-a",
-            type: "tool",
-            tool: "task",
-            state: {
-              status: "running",
-              input: {
-                subagent_type: "build",
-                prompt: "Inspect repo",
-                description: "Starting A",
-              },
-              metadata: {
-                externalSessionId: "child-a",
-              },
-            },
-          },
-        },
-      } as unknown as Event,
-    ]);
-
-    const assistantPartEvents = emitted.filter(
-      (event): event is Extract<AgentEvent, { type: "assistant_part" }> =>
-        event.type === "assistant_part",
-    );
-    const subagentParts = assistantPartEvents
-      .map((event): Extract<AgentEvent, { type: "assistant_part" }>["part"] => event.part)
-      .filter(
-        (
-          part,
-        ): part is Extract<
-          Extract<AgentEvent, { type: "assistant_part" }>["part"],
-          { kind: "subagent" }
-        > => part.kind === "subagent",
-      );
-
-    expect(subagentParts).toHaveLength(2);
-    expect(subagentParts.map((part) => part.correlationKey)).toEqual([
-      "part:assistant-task-tool-running:subtask-a",
-      "part:assistant-task-tool-running:subtask-a",
-    ]);
-    expect(subagentParts.map((part) => part.externalSessionId)).toEqual([undefined, "child-a"]);
-    expect(subagentParts.map((part) => part.agent)).toEqual(["build", "build"]);
-  });
-
-  test("clears pending subagent queues once a running task tool update gains a session id", async () => {
-    const { sessionRecord } = await runEventStreamWithSession([
-      assistantRoleEvent("assistant-task-tool-running"),
-      makeAssistantSubtaskPartUpdatedEvent({
-        messageId: "assistant-task-tool-running",
-        partId: "subtask-a",
-        agent: "build",
-        prompt: "Inspect repo",
-        description: "Starting A",
-      }),
-      {
-        type: "message.part.updated",
-        properties: {
-          part: {
-            id: "tool-a",
-            sessionID: "external-session-1",
-            messageID: "assistant-task-tool-running",
-            callID: "call-a",
-            type: "tool",
-            tool: "task",
-            state: {
-              status: "running",
-              input: {
-                subagent_type: "build",
-                prompt: "Inspect repo",
-                description: "Starting A",
-              },
-              metadata: {
-                externalSessionId: "child-a",
-              },
-            },
-          },
-        },
-      } as unknown as Event,
-    ]);
-
-    expect(sessionRecord.subagentCorrelationKeyByExternalSessionId.get("child-a")).toBe(
-      "part:assistant-task-tool-running:subtask-a",
-    );
-    expect(sessionRecord.pendingSubagentCorrelationKeys).toEqual([]);
-    expect(sessionRecord.pendingSubagentCorrelationKeysBySignature.size).toBe(0);
-  });
-
-  test("binds sibling child sessions correctly when session.created arrives out of order", async () => {
-    const { emitted } = await runEventStreamWithSession([
-      assistantRoleEvent("assistant-subagent-created-order"),
-      makeAssistantSubtaskPartUpdatedEvent({
-        messageId: "assistant-subagent-created-order",
-        partId: "subtask-a",
-        agent: "build",
-        prompt: "Inspect repo",
-        description: "Starting A",
-      }),
-      makeAssistantSubtaskPartUpdatedEvent({
-        messageId: "assistant-subagent-created-order",
-        partId: "subtask-b",
-        agent: "build",
-        prompt: "Inspect repo",
-        description: "Starting B",
-      }),
-      {
-        type: "session.created",
-        properties: {
-          info: {
-            id: "child-b",
-            parentID: "external-session-1",
-            time: { created: Date.parse("2026-02-22T12:00:12.000Z") },
-          },
-        },
-      } as unknown as Event,
-      {
-        type: "session.created",
-        properties: {
-          info: {
-            id: "child-a",
-            parentID: "external-session-1",
-            time: { created: Date.parse("2026-02-22T12:00:10.000Z") },
-          },
-        },
-      } as unknown as Event,
-      {
-        type: "message.part.updated",
-        properties: {
-          part: {
-            id: "tool-a",
-            sessionID: "external-session-1",
-            messageID: "assistant-subagent-created-order",
-            callID: "call-a",
-            type: "tool",
-            tool: "delegate",
-            state: {
-              status: "completed",
-              input: {
-                agent: "build",
-                prompt: "Inspect repo",
-              },
-              output: {
-                result: "Finished A",
-                externalSessionId: "child-a",
-              },
-            },
-          },
-        },
-      } as unknown as Event,
-      {
-        type: "message.part.updated",
-        properties: {
-          part: {
-            id: "tool-b",
-            sessionID: "external-session-1",
-            messageID: "assistant-subagent-created-order",
-            callID: "call-b",
-            type: "tool",
-            tool: "delegate",
-            state: {
-              status: "completed",
-              input: {
-                agent: "build",
-                prompt: "Inspect repo",
-              },
-              output: {
-                result: "Finished B",
-                externalSessionId: "child-b",
-              },
-            },
-          },
-        },
-      } as unknown as Event,
-    ]);
-
-    const completedSubagentParts = emitted.filter(
-      (event): event is Extract<AgentEvent, { type: "assistant_part" }> =>
-        event.type === "assistant_part" &&
-        event.part.kind === "subagent" &&
-        event.part.status === "completed",
-    );
-
-    expect(
-      completedSubagentParts.map(
-        (event) => (event.part as { correlationKey: string }).correlationKey,
-      ),
-    ).toEqual([
-      "part:assistant-subagent-created-order:subtask-a",
-      "part:assistant-subagent-created-order:subtask-b",
-    ]);
-    expect(
-      completedSubagentParts.map(
-        (event) => (event.part as { externalSessionId?: string }).externalSessionId,
-      ),
-    ).toEqual(["child-a", "child-b"]);
-  });
-
-  test("emits a linked subagent update when child session creation binds a running card", async () => {
-    const { emitted } = await runEventStreamWithSession([
-      assistantRoleEvent("assistant-subagent-session-created"),
-      makeAssistantSubtaskPartUpdatedEvent({
-        messageId: "assistant-subagent-session-created",
-        partId: "subtask-a",
-        agent: "build",
-        prompt: "Inspect repo",
-        description: "Starting A",
-      }),
-      makeChildSessionCreatedEvent({
-        childSessionId: "external-child-session",
-        parentExternalSessionId: "external-session-1",
-      }),
-    ]);
-
-    const subagentParts = emitted
-      .filter(
-        (event): event is Extract<AgentEvent, { type: "assistant_part" }> =>
-          event.type === "assistant_part" && event.part.kind === "subagent",
-      )
-      .map((event) => event.part as Extract<AgentEvent, { type: "assistant_part" }>["part"]);
-
-    expect(subagentParts).toHaveLength(2);
-    expect(subagentParts.map((part) => part.correlationKey)).toEqual([
-      "part:assistant-subagent-session-created:subtask-a",
-      "part:assistant-subagent-session-created:subtask-a",
-    ]);
-    expect(subagentParts.map((part) => part.externalSessionId)).toEqual([
-      undefined,
-      "external-child-session",
-    ]);
-  });
-
-  test("defers ambiguous task tool updates until child sessions bind to spawned cards", async () => {
-    const { emitted } = await runEventStreamWithSession([
-      assistantRoleEvent("assistant-subagent-ambiguous-deferred"),
-      makeAssistantSubtaskPartUpdatedEvent({
-        messageId: "assistant-subagent-ambiguous-deferred",
-        partId: "subtask-a",
-        agent: "build",
-        prompt: "Inspect repo",
-        description: "Starting A",
-      }),
-      makeAssistantSubtaskPartUpdatedEvent({
-        messageId: "assistant-subagent-ambiguous-deferred",
-        partId: "subtask-b",
-        agent: "build",
-        prompt: "Inspect repo",
-        description: "Starting B",
-      }),
-      {
-        type: "message.part.updated",
-        properties: {
-          part: {
-            id: "tool-b",
-            sessionID: "external-session-1",
-            messageID: "assistant-subagent-ambiguous-deferred",
-            callID: "call-b",
-            type: "tool",
-            tool: "task",
-            state: {
-              status: "completed",
-              metadata: {
-                agent: "build",
-                prompt: "Inspect repo",
-                externalSessionId: "child-b",
-              },
-            },
-          },
-        },
-      } as unknown as Event,
-      {
-        type: "session.created",
-        properties: {
-          info: {
-            id: "child-b",
-            parentID: "external-session-1",
-            time: { created: Date.parse("2026-02-22T12:00:12.000Z") },
-          },
-        },
-      } as unknown as Event,
-      {
-        type: "session.created",
-        properties: {
-          info: {
-            id: "child-a",
-            parentID: "external-session-1",
-            time: { created: Date.parse("2026-02-22T12:00:10.000Z") },
-          },
-        },
-      } as unknown as Event,
-    ]);
-
-    const subagentParts = emitted.filter(
-      (event): event is Extract<AgentEvent, { type: "assistant_part" }> =>
-        event.type === "assistant_part" && event.part.kind === "subagent",
-    );
-
-    expect(subagentParts).toHaveLength(4);
-    expect(
-      subagentParts.map((event) => (event.part as { correlationKey: string }).correlationKey),
-    ).toEqual([
-      "part:assistant-subagent-ambiguous-deferred:subtask-a",
-      "part:assistant-subagent-ambiguous-deferred:subtask-b",
-      "part:assistant-subagent-ambiguous-deferred:subtask-a",
-      "part:assistant-subagent-ambiguous-deferred:subtask-b",
-    ]);
-    expect(subagentParts.map((event) => (event.part as { status: string }).status)).toEqual([
-      "running",
-      "running",
-      "running",
-      "completed",
-    ]);
-    expect(
-      subagentParts.map(
-        (event) => (event.part as { externalSessionId?: string }).externalSessionId,
-      ),
-    ).toEqual([undefined, undefined, "child-a", "child-b"]);
   });
 
   test("normalizes todo.updated and ignores unrelated sessions", async () => {
