+++ packages/adapters-opencode-sdk/src/event-stream/message-events.ts
@@ -7,10 +7,7 @@ import {
 import { handleMessageUpdatedEvent } from "./message-events/updated";
 import type { EventStreamRuntime } from "./shared";
 
-export {
-  emitSubagentSessionLinkUpdate,
-  flushPendingSubagentPartEmissionsForSession,
-} from "./message-events/assistant";
+export { emitSubagentPartsForSession } from "./message-events/assistant";
 export { reconcileUserMessageQueuedStates } from "./message-events/user";
 
 export const handleMessageEvent = (event: Event, runtime: EventStreamRuntime): boolean => {
