+++ packages/adapters-opencode-sdk/src/event-stream/session-events.ts
@@ -3,11 +3,7 @@ import type { AgentEvent } from "@openducktor/core";
 import { toAgentApprovalRequestFromOpenCodePermission } from "../approval-translation";
 import { readStringProp } from "../guards";
 import { normalizeTodoList } from "../todo-normalizers";
-import {
-  emitSubagentSessionLinkUpdate,
-  flushPendingSubagentPartEmissionsForSession,
-  reconcileUserMessageQueuedStates,
-} from "./message-events";
+import { emitSubagentPartsForSession, reconcileUserMessageQueuedStates } from "./message-events";
 import {
   parsePermissionAsked,
   parseQuestionAsked,
@@ -263,15 +259,7 @@ const bindChildSessionCorrelation = (event: Event, runtime: EventStreamRuntime):
     normalizedChildExternalSessionId,
   );
   if (existingCorrelationKey && !existingCorrelationKey.startsWith("session:")) {
-    const hasPendingPartEmissions = Boolean(
-      runtime.pendingSubagentPartEmissionsByExternalSessionId.get(normalizedChildExternalSessionId)
-        ?.length,
-    );
-    if (hasPendingPartEmissions) {
-      flushPendingSubagentPartEmissionsForSession(runtime, normalizedChildExternalSessionId);
-    } else {
-      emitSubagentSessionLinkUpdate(runtime, normalizedChildExternalSessionId);
-    }
+    emitSubagentPartsForSession(runtime, normalizedChildExternalSessionId);
     flushPendingSubagentInputEventsForSession(runtime, normalizedChildExternalSessionId);
     return true;
   }
@@ -310,14 +298,7 @@ const bindChildSessionCorrelation = (event: Event, runtime: EventStreamRuntime):
     runtime.subagentCorrelationKeyByExternalSessionId.set(externalSessionId, nextCorrelationKey);
     runtime.pendingSubagentSessionsByExternalSessionId.delete(externalSessionId);
     removePendingSubagentCorrelationKey(runtime, nextCorrelationKey);
-    const hasPendingPartEmissions = Boolean(
-      runtime.pendingSubagentPartEmissionsByExternalSessionId.get(externalSessionId)?.length,
-    );
-    if (hasPendingPartEmissions) {
-      flushPendingSubagentPartEmissionsForSession(runtime, externalSessionId);
-    } else {
-      emitSubagentSessionLinkUpdate(runtime, externalSessionId);
-    }
+    emitSubagentPartsForSession(runtime, externalSessionId);
     flushPendingSubagentInputEventsForSession(runtime, externalSessionId);
   }
   return true;
